pyproject.toml
[tool.poetry]
name = "microboss"
version = "0.1.0"
description = "An AI agent system that decomposes complex tasks"
authors = ["Microboss Contributors <example@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
tqdm = "^4.66.1"  # For progress bars
python-dotenv = "^1.0.0"  # For loading environment variables
flask = "^2.3.0"  # Web framework
flask-socketio = "^5.3.0"  # Real-time communication
markdown = "^3.5.0"  # For converting markdown to HTML
pygments = "^2.15.0"  # For code highlighting
plotly = "^5.15.0"  # For visualization
dash = "^2.11.0"  # For interactive visualizations
anthropic = "^0.23.0"  # Updated to the latest version
openai = "^1.15.0"  # OpenAI API for GPT fallback
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
microboss = "microboss.cli:main"
microboss-web = "microboss.web.app:main"
